Switch to using a trait with the validation methods to provide easier sharing between distribution classes
